+++ crates/system-env/src/lib.rs
@@ -1,6 +1,7 @@
 mod deps;
 mod env;
 mod error;
+mod helpers;
 mod pm;
 mod pm_vendor;
 mod system;
@@ -8,6 +9,7 @@ mod system;
 pub use deps::*;
 pub use env::*;
 pub use error::*;
+pub use helpers::*;
 pub use pm::*;
 pub use pm_vendor::*;
 pub use system::*;
